requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-crispy-forms==1.13.0
django-phonenumber-field==6.0.0
phonenumbers==8.12.39
Pillow==8.4.0
python-decouple==3.5
sqlparse==0.4.2
